Implement patterned rendering
This is the current default, but can be changed by
changing the pattern_grid_size value in Material.

Note that it doesn't draw a proper spherical test pattern,
as that would require me to special case shading of spheres.
Instead, we use the same way of rendering patterns for all
primitives, which looks a bit funny on a sphere, but
altogether not too bad.